[better-sequence] - Created smarter AsyncSequence - TT
